+++ gateway-js/src/__tests__/gateway/reporting.test.ts
@@ -1,11 +1,8 @@
 import { gunzipSync } from 'zlib';
 import nock from 'nock';
-import gql from 'graphql-tag';
 import { buildSubgraphSchema } from '@apollo/subgraph';
 import { ApolloServer } from 'apollo-server';
 import { ApolloServerPluginUsageReporting } from 'apollo-server-core';
-import { execute, toPromise } from 'apollo-link';
-import { createHttpLink } from 'apollo-link-http';
 import fetch from 'node-fetch';
 import { ApolloGateway } from '../..';
 import { Plugin, Config, Refs } from 'pretty-format';
@@ -141,7 +138,7 @@ describe('reporting', () => {
   });
 
   it(`queries three services`, async () => {
-    const query = gql`
+    const query = `#graphql
       query {
         me {
           name {
@@ -155,12 +152,15 @@ describe('reporting', () => {
       }
     `;
 
-    const result = await toPromise(
-      execute(createHttpLink({ uri: gatewayUrl, fetch: fetch as any }), {
-        query,
-      }),
-    );
-    expect(result).toMatchInlineSnapshot(`
+    const result = await fetch(gatewayUrl, {
+      method: 'POST',
+      headers: {
+        'Content-Type': 'application/json',
+      },
+      body: JSON.stringify({ query }),
+    });
+    const body = await result.json();
+    expect(body).toMatchInlineSnapshot(`
       Object {
         "data": Object {
           "me": Object {
@@ -267,7 +267,6 @@ describe('reporting', () => {
                 },
                 "clientName": "",
                 "clientVersion": "",
-                "details": Object {},
                 "durationNs": 12345,
                 "endTime": Object {
                   "nanos": 123000000,