+++ query-planner-js/src/composedSchema/removeInaccessibleElements.ts
@@ -15,7 +15,7 @@ import {
   GraphQLError,
   GraphQLCompositeType,
 } from 'graphql';
-import { transformSchema } from 'apollo-graphql';
+import { transformSchema } from '../schema-helper';
 
 export function removeInaccessibleElements(
   schema: GraphQLSchema,
@@ -130,7 +130,7 @@ export function removeInaccessibleElements(
     return newFieldMapConfig;
   }
 
-  function removeInaccessibleTypes<T extends GraphQLNamedType>(types: T[]) {
+  function removeInaccessibleTypes<T extends GraphQLNamedType>(types: readonly T[]) {
     return types.filter((type) => !typesToRemove.has(type));
   }
 }